# docs/swagger.yaml
definitions:
  middle.ResCode:
    enum:
    - 200
    - 201
    - 202
    - 203
    - 204
    - 205
    - 206
    - 207
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeInvalidParam
    - CodeUserExist
    - CodeUserNotExist
    - CodeInvalidPassword
    - CodeServerBusy
    - CodeNeedLogin
    - CodeInvalidToken
  middle.ResponseData:
    properties:
      code:
        $ref: '#/definitions/middle.ResCode'
      data: {}
      msg: {}
    type: object
...
  service.ServiceCreate:
    properties:
      container_port:
        type: integer
      label:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      namespace:
        type: string
      node_port:
        type: integer
      port:
        type: integer
      type:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/k8s/Secret/del:
    delete:
      consumes:
      - application/json
      description: 删除Secret
      operationId: /api/k8s/Secret/del
      parameters:
      - description: Secret名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "删除成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 删除Secret
      tags:
      - Secret
  /api/k8s/Secret/detail:
    get:
      consumes:
      - application/json
      description: 获取Secret详情
      operationId: /api/k8s/Secret/detail
      parameters:
      - description: Secret名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data":v1.Deployment }'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 获取Secret详情
      tags:
      - Secret
  /api/k8s/Secret/list:
    get:
      consumes:
      - application/json
      description: 查看Secret列表
      operationId: /api/k8s/Secret/list
      parameters:
      - description: 过滤
        in: query
        name: filter_name
        type: string
      - description: 命名空间
        in: query
        name: namespace
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 分页限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": }'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 查看Secret列表
      tags:
      - Secret
...
  /api/k8s/secret/update:
    put:
      consumes:
      - application/json
      description: 更新Secret
      operationId: /api/k8s/secret/update
      parameters:
      - description: 无状态控制器名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: 更新内容
        in: query
        name: content
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "更新成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 更新Secret
      tags:
      - Secret
  /api/k8s/service/create:
    post:
      consumes:
      - application/json
      description: 创建service
      operationId: /api/k8s/service/create
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.ServiceCreate'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "创建成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 创建service
      tags:
      - service
  /api/k8s/service/del:
    delete:
      consumes:
      - application/json
      description: 删除service
      operationId: /api/k8s/service/del
      parameters:
      - description: service名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "删除成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 删除service
      tags:
      - service
  /api/k8s/service/detail:
    get:
      consumes:
      - application/json
      description: 获取service详情
      operationId: /api/k8s/service/detail
      parameters:
      - description: service名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data":v1.Deployment }'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 获取service详情
      tags:
      - service
  /api/k8s/service/list:
    get:
      consumes:
      - application/json
      description: 查看service列表
      operationId: /api/k8s/service/list
      parameters:
      - description: 过滤
        in: query
        name: filter_name
        type: string
      - description: 命名空间
        in: query
        name: namespace
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 分页限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": }'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 查看service列表
      tags:
      - service
...
  /api/k8s/service/update:
    put:
      consumes:
      - application/json
      description: 更新service
      operationId: /api/k8s/service/update
      parameters:
      - description: service名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: 更新内容
        in: query
        name: content
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "更新成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 更新service
      tags:
      - service
...
swagger: "2.0"